File name change and moving useEffect below useState decaration.
